requirements.txt
azure-core==1.26.2
azure-storage-blob==12.14.1
beautifulsoup4==4.8.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
cryptography==39.0.0
html-parser==0.2
idna==2.8
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
ply==3.11
pycparser==2.21
python-dateutil==2.8.2
requests==2.22.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.25.11
